conf: improve capability handling

+++ src/lxc/macro.h
@@ -770,19 +770,19 @@ typedef long long unsigned int llu;
 #define DIV_ROUND_UP(n, d) (((n) + (d)-1) / (d))
 #define BITS_TO_LONGS(nr) DIV_ROUND_UP(nr, NBITS)
 
-static inline void set_bit(unsigned bit, uint32_t *bitarr)
+static inline void set_bit(__u32 bit, __u32 *bitarr)
 {
-	bitarr[bit / NBITS] |= (1 << (bit % NBITS));
+	bitarr[bit / NBITS] |= ((__u32)1 << (bit % NBITS));
 }
 
-static inline void clear_bit(unsigned bit, uint32_t *bitarr)
+static inline void clear_bit(__u32 bit, __u32 *bitarr)
 {
-	bitarr[bit / NBITS] &= ~(1 << (bit % NBITS));
+	bitarr[bit / NBITS] &= ~((__u32)1 << (bit % NBITS));
 }
 
-static inline bool is_set(unsigned bit, uint32_t *bitarr)
+static inline bool is_set(__u32 bit, __u32 *bitarr)
 {
-	return (bitarr[bit / NBITS] & (1 << (bit % NBITS))) != 0;
+	return (bitarr[bit / NBITS] & ((__u32)1 << (bit % NBITS))) != 0;
 }
 
 #endif /* __LXC_MACRO_H */
